#55 [FEAT] 학교 이메일 인증 로직 및 회원가입 정보 제출 로직을 훅 내부로 이동
