+++ src/pages/SignUp/UniversityApproval.jsx
@@ -1,16 +1,15 @@
 import React, { useState } from 'react';
-import { submitSignUpInfo } from '../../apis/signUp';
 import { ReactComponent as EmailIcon } from '../../assets/outgoingMail.svg';
 import { ReactComponent as FileIcon } from '../../assets/uploadFile.svg';
 import AutnButton from '../../components/Button/AuthButton';
 import OrangeButton from '../../components/Button/OrangeButton';
 import WhiteButton from '../../components/Button/WhiteButton';
-import { useSignUpContext } from '../../contexts/SignUpContext';
+import useSignUp from '../../hooks/useSignUp';
 import EmailApproval from './EmailApproval';
 import FileApproval from './FileApproval';
 
 export default function UniversityApproval({ moveStep }) {
-  const { signUpInfo } = useSignUpContext();
+  const { submitSignUpInfo } = useSignUp();
   const [file, setFile] = useState();
   const [isSelected, setIsSleceted] = useState('');
   const changeMenu = (event) => setIsSleceted(event.currentTarget.name);
@@ -59,7 +58,7 @@ export default function UniversityApproval({ moveStep }) {
           <div className='w-[148px] h-[51px] text-[18px]'>
             <OrangeButton
               text='다음'
-              onClick={() => submitSignUpInfo(file, signUpInfo, moveStep)}
+              onClick={() => submitSignUpInfo(file, moveStep)}
               disabled={!isPassed}
             />
           </div>
